DRF-shop-project/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djoser==2.2.2
pillow==10.2.0

